bugfix on macOS
