+++ FullScreenTranslator/Views/Common/ScaleButtonStyle.swift
@@ -1,0 +1,10 @@
+import SwiftUI
+
+/// ボタンがタップされたときにアニメーションするボタンスタイル
+struct ScaleButtonStyle: ButtonStyle {
+  func makeBody(configuration: Configuration) -> some View {
+    configuration.label
+      .scaleEffect(configuration.isPressed ? 0.95 : 1)
+      .animation(.easeInOut(duration: 0.1), value: configuration.isPressed)
+  }
+}
