+++ FullScreenTranslator/Views/macOS/macOSAppScene.swift
@@ -13,6 +13,7 @@ import SwiftUI
           translated: viewModel.translator.translated
         )
         .frame(maxWidth: .infinity, maxHeight: .infinity)
+        .translatorViewModifiers(viewModel: viewModel)
       }
       .windowStyle(HiddenTitleBarWindowStyle())
 
@@ -43,7 +44,6 @@ import SwiftUI
               Image(systemName: "translate")
             }
           }
-          .translatorViewModifiers(viewModel: viewModel)
         }
       )
     }
